[maint] Move loading to common
