+++ mspelling/lib/views/setup_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mspelling/controllers/setup_controller.dart';
-import 'package:mspelling/views/loading_view.dart';
+import 'package:mspelling/common/loading_view.dart';
 
 class SetupView extends StatelessWidget {
   SetupView({super.key});
